+++ scripts/example-usebench.sh
@@ -9,8 +9,11 @@ set +a
 
 if [[ "$1" == "--build" ]]; then
   docker rm -f useagent-turbo-test 2>/dev/null
-  docker image rm -f useagent-turbo:dev 2>/dev/null
-    DOCKER_BUILDKIT=1 docker build --build-arg BASE_IMAGE=usebench.sweb.eval.x86_64.django__django-10914 --ssh default -t useagent-turbo:dev .
+  if [[ "$2" == "--rm-image" ]]; then
+    #This will delete the image, but you will loose caching.
+    docker image rm -f useagent-turbo:dev 2>/dev/null
+  fi
+  DOCKER_BUILDKIT=1 docker build --build-arg BASE_IMAGE=usebench.sweb.eval.x86_64.django__django-10914 --ssh default -t useagent-turbo:dev .
 fi
 
 docker run --rm \
